Made crumbs clickable to navigate to previous groups
