+++ src/h5web/breadcrumbs/BreadcrumbsBar.module.css
@@ -21,6 +21,15 @@
   white-space: nowrap;
 }
 
+.crumbButton {
+  composes: btn-clean from global;
+  display: flex;
+}
+
+.crumbButton:hover {
+  text-decoration: underline;
+}
+
 .crumb {
   margin: 0 0.125rem 0.125rem;
   overflow: hidden;
